Command line arguments being implemented
Application now accepts a single parameter representing the number of random
numbers to generate. The string is parsed and converted into a numeric value,
essentially reimplementing `atoi` functionality from the C standard library.

The program then loops 'n' times, generating and printing the specified number
of random numbers. If no argument is passed in, the default is 1.